moved site level schema definition to app level
